Refactor - remove 'base' from file names
